перенес логику отрисовки листов и карточек в новый класс пейдж контроллер, некорректно работает обработчик по карточке
