+++ src/components/abstract.js
@@ -4,7 +4,7 @@ class Abstract {
   constructor() {
     if (new.target === Abstract) {
       throw new Error(`Can't instantiate Abstract, only concrete one.`);
-    };
+    }
     this._element = null;
   }
