+++ src/main.js
@@ -2,11 +2,9 @@
 import {Profile} from './components/profile.js';
 import {Navigation} from './components/navigation.js';
 import {Sort} from './components/sort.js';
-import {FilmList, getEntitiesForRender, getIsMaxFilms} from './components/film-list.js';
-import {Film, getRandomFilmEntity} from './components/film.js';
-import {FilmDetail} from './components/film-detail.js';
-import {ShowMoreButton} from './components/show-more-button.js';
-import {Keycode, sortArrWithObjByKey} from './components/utils.js';
+import {getRandomFilmEntity} from './components/film.js';
+
+import PageController from './components/page-controller.js';
 
 const ClassName = {
   HEADER: `header`,
@@ -77,8 +75,6 @@ const FilmListConfig = {
   }
 };
 
-let currentFilmDetail = null;
-
 const filmEntites = new Array(FilmListConfig.General.Count.MAX).fill(``).map((item, i) => {
   return getRandomFilmEntity(i);
 });
@@ -125,65 +121,6 @@ const getRaiting = (filmsCount) => {
   return raiting;
 };
 
-const onFilmElementClick = (evt) => {
-  evt.preventDefault();
-  const element = evt.currentTarget;
-  const currentEntity = filmEntites[element.dataset.id];
-  console.log(evt.target.classList.contains(ClassName.FILM_POSTER || ClassName.FILM_TITLE || ClassName.FILM_COMMENT_COUNT))
-
-  if (evt.target.classList.contains(ClassName.FILM_TITLE || ClassName.FILM_COMMENT_COUNT)) {
-
-    currentFilmDetail = new FilmDetail(currentEntity);
-    currentFilmDetail.renderElement(document.body);
-    currentFilmDetail.addClickHandlerOnElement(onFilmDetailElementClick);
-    document.addEventListener(`keydown`, onDocumentKeydown);
-  }
-};
-
-const onFilmDetailElementClick = (evt) => {
-  evt.preventDefault();
-  if (evt.target.classList.contains(ClassName.FILM_DETAIL_CLOSE_BUTTON)) {
-    currentFilmDetail.removeElement();
-    document.removeEventListener(`keydown`, onDocumentKeydown);
-  }
-};
-
-const onShowMoreButtonElementClick = (evt) => {
-  evt.preventDefault();
-  renderFilmsInGeneralList();
-  if (getIsMaxFilms()) {
-    showMoreButton.removeElement();
-  }
-};
-
-const onDocumentKeydown = (evt) => {
-  evt.preventDefault();
-  if (evt.keyCode === Keycode.ESC) {
-    currentFilmDetail.removeElement();
-    document.removeEventListener(`keydown`, onDocumentKeydown);
-  }
-};
-
-const renderFilms = (config, filmList) => {
-  const entities = sortArrWithObjByKey(filmEntites, config.SORT_PROPERTY);
-
-  getEntitiesForRender(entities, config).forEach((item) => {
-    const film = new Film(item);
-    film.renderElement(filmList.getContainerElement());
-    film.addClickHandlerOnElement(onFilmElementClick);
-  });
-};
-
-const renderFilmsInGeneralList = () => {
-  renderFilms(FilmListConfig.General, generalFilmList);
-  generalFilmList.hideEmptyElement();
-};
-
-const renderFilmsInExtraList = (config, filmList) => {
-  renderFilms(config, filmList);
-  filmList.hideEmptyElement();
-};
-
 const headerElement = document.querySelector(`.${ClassName.HEADER}`);
 const mainElement = document.querySelector(`.${ClassName.MAIN}`);
 const contentElement = mainElement.querySelector(`.${ClassName.CONTENT}`);
@@ -196,26 +133,10 @@ sort.renderElement(mainElement, `afterbegin`);
 const navigation = new Navigation(statistic);
 navigation.renderElement(mainElement, `afterbegin`);
 
-if (filmEntites.length === 0) {
-  const noDataFilmList = new FilmList(FilmListConfig.NoData);
-  noDataFilmList.renderElement(contentElement);
-}
 
-const generalFilmList = new FilmList(FilmListConfig.General);
-generalFilmList.renderElement(contentElement);
-renderFilmsInGeneralList();
+const pageController = new PageController(contentElement);
+pageController.render(filmEntites);
 
-const topRatedFilmList = new FilmList(FilmListConfig.TopRated);
-topRatedFilmList.renderElement(contentElement);
-renderFilmsInExtraList(FilmListConfig.TopRated, topRatedFilmList);
-
-const mostCommentedFilmList = new FilmList(FilmListConfig.MostCommented);
-mostCommentedFilmList.renderElement(contentElement);
-renderFilmsInExtraList(FilmListConfig.MostCommented, mostCommentedFilmList);
-
-const showMoreButton = new ShowMoreButton();
-showMoreButton.renderElement(generalFilmList.getElement());
-showMoreButton.addClickHandlerOnElement(onShowMoreButtonElementClick);
 
 const setFilmsCount = () => {
   const element = document.querySelector(`.${ClassName.FILMS_COUNT}`);
@@ -223,4 +144,4 @@ const setFilmsCount = () => {
 };
 setFilmsCount();
 
-export {FilmListConfig, renderFilmsInGeneralList};
+export {FilmListConfig, ClassName};
